refactor: 구분자 추출 클래스 분리
1. Separator 클래스 추가
